Workaround for FilteredDocumentType
There is a bug in the data model where the metadata lookup for
FilteredDocumentType a) contains the fully qualified path (unnecessary),
and b) does not actually map to the ContentType names. This is worked
around with an IndexOf and a switch statement. This should be removed
once the bug is fixed.