Hack: Added additional animation
When skipping over accordion items, the previous animation doesn't
always complete. This may be due to competition from other animations
occuring on the page (?). To ensure that final step is completed, I am
rerunning the animation after 200ms. If the original animation
succeeded, this shouldn't be necessary; otherwise, it should "read" as
the final step in that transition. This is not elegant, but seems to
address some inconsistent behavior with `scrollTop`.