Moved single-page templates to `/ContentTypes`
With the new `TopicViewEngine`, content type views can be placed in the `~/Views/ContentTypes` folder. This is useful for stand-alone templates that aren't "overloaded" with multiple views, as it helps flatten the hierarchy, instead of having a 1:1 relationship between folders and views. Content types with multiple views continue to be best served by being placed in their own directory (e.g., `~/Views/ContentList/*.cshtml`).