Implemented `FallbackController`
Implemented a `FallbackController` which will be used as the default if no valid controller is found. This allows a 404 to be thrown instead. (Technically, this could be handled as a 500 exception, but it's a fairly common error that generally means a mistyped, legacy, or otherwise invalid URL so it's preferable to treat it as a 404).