Renamed to `TopicEntityViewModel`
To avoid an unnecessary naming conflict, renamed `TopicViewModel` to `LegacyTopicViewModel`. This also helps signal that this is intended to work with the `Topic` entity, and not the `TopicViewModel`. By the time this refactoring is complete, we should be able to remove most references to the `TopicEntityViewModel` (an exception might be the `Sitemap.cshtml`).