Added `ModulePageTopic`
In order for modules to be treated as cards, they need to implement the `ICardViewModel` interface. Currently, modules are `Page` content types. This will require that they be changed to a new `ModulePage` content type, which is simply a wrapper for `Page` (it can have the same data model, except it'll add `ThumbnailImage`).