Fixed logic for headers
The headers should be linked if the user is on desktop or if there are two or fewer children—as, in that case, there won't be any children displayed, and the link should be active. This logic apparently got flipped. (I believe, previously, there was a `hasChildren` variable that was marked as `!hasChildren` in this case, but the logic wasn't flipped to account for that when the variable was removed.)