Implemented `LayoutControllerBase`
Much of the logic for `LayoutController` has now been centralized in the `LayoutControllerBase` as part of `Ignia.Topics.Web.Mvc`. This implements that change by deriving from `LayoutControllerBase` and using the newly centralized `NavigationViewModel`. In addition, this utilizes a GoldSim-specific version of `NavigationViewModel` so as to support `BannerUrl` (which isn't supported by the global Ignia version).