Implemented `ErrorControllerBase`
The `ErrorController` logic has been centralized as part of the `ErrorControllerBase` class in `Ignia.Topics.Web.Mvc`. This class is generic to allow any view models to be used. Since MVC doesn't route effectively with generics, this class must be derived from, and supplied with the appropriate view model class (in GoldSim's case, the `PageTopicViewModel` works fine).