Centralize `RedirectController`
Much of the logic for the `RedirectController` is centralized as part of `Ignia.Topics.Web.Mvc`. That said, GoldSim has additional logic that is custom to GoldSim, thus requiring a custom `RedirectController` with additional actions. This introduces ambiguities in the naming conventions, thus additionally necessitating that the `namespaces` attribute be added to the corresponding routes.