Established `applicationInitialization`
This should allow the application to automatically initialize before responding to any requests, thus allowing the app to warm up the cache independent of the first request. This means that when the app resets it will immediately start loading the cache, instead of waiting for the first web request—and, thus, making that first request wait for the cache to finish loading.