Extended application initialization
Ensured that the application initialization hits an instance of a page representing every view, including the major forms. That will act as a temporary solution until we're able to fully-enable precompilation of views on publish (which currently isn't practical due to the dependencies on some legacy WebForm components, such as the forms and editor). In addition, introduced the `app_initializing.htm` so that a friendly message is displayed should any customers be accessing the site during this preload period.