Made `GetConfigurationSetting()` multi-purpose
Instead of using `GetConfigurationSetting()` _exclusively_ for puling from the `AppSettings`, it now handles _all_ configuration sources. In addition, instead of ignoring and overriding the locally configured instance values set on the properties, it will always prioritize those, if they're set. This greatly reduces the amount of code required, and centralizes the logic for retrieving configuration values.