Included port in email receipt
The port is optional in the production environment, since it runs on the default HTTP/S ports. In the development environment, however, the server will usually be running on a custom port. As such, this needs to be included in the URL for `SendWebPage`, otherwise it will receive a 404 when sending.