Injected null check for `GetCurrentTopic()`
In some circumstances, `GetCurrentTopic()` will be `null`. Typically, that pertains to a 404 page. And, in those situations, the controller will redirect to `NotFound`. But before it can get there, it needs to make it through this code. As such, `GetCurrentTopic()` needs a null check on it to ensure it doesn't fail if it turns `null`.