Updated to .NET 8.0
There was an issue on my machine of locating .NET 4.7.1, which is required to target the current version of OnTopic. That shouldn'tnormally be an issue, and I'm looking into it, but in the meanwhile I just updated to .NET 4.8, which recently relieased, and that resolves the issue. This is also the final release of .NET Framework.