Fixed Range Overload
According to the documentation, the `ErrorMessage` property shouldn't need to be defined in .NET Core 3.0. Since the site hasn't been migrated to .NET Core 3.0, however, the property still must be explicitly stated for now. We _may_ be able to remove this later, once we target .NET Core 3.0. For now, though, it allows the project to compile and won't hurt anything to have defined explicitly, as we do elsewhere.