Upgrade to .NET Framework 4.8
In order to target the latest version of the *OnTopic* library, GoldSim needs to be on _at least_ *.NET Framework 4.7.2*. That said, since *.NET Framework 4.8* just released, and there aren't any known compatibility issues, we're leapfrogging *4.7.2* to prevent the need to upgrade later. (Though, of course, we'll need to _migrate_ to *.NET Core 3.0* at some point.) This also updates the *C# compiler* version from 1.0.8.0 to 2.0.1.0.