Updated *OnTopic* to `DotNetStandard` prerelease for testing
The `DotNetStandard` branch (and derivatives) has not yet been merged into the `Master` branch or released. That said, it is currently deployed to `develop` as a prerelease for testing. To confirm that it works with *GoldSim*, this update upgrades the NuGet depenencies for testing purposes. If this succeed, this will need to be updated to the final release once it's been merged into `Master`.