Updated to use new `HierarchicalTopicMappingService`
The new `HierarchicalTopicMappingService` encompasses a lot of logic previously handled by the `LayoutController`, and helps interrmediate the previous differences between the `CachedLayoutControllerBase` and the `LayoutControllerBase` by treating that as a lifestyle decision. This is a breaking change, and requires updates to the concrete `LayoutController` as well as the _composition root_ (i.e., the `GoldSimControllerFactory`).