Updated to `alpha.257` build of *OnTopic*
Initially, the `DotNetStandard` branch was operating off of the `dotnetstandard0001` build, which was triggered directly from the `DotNetStandard` branch. That branch has since been merged with `Develop`, alongside some other ancillary changes, and is considered a release candidate for the *.NET Standard* version of *OnTopic*. This update ensures *GoldSim* is being tested against this latest build.