Migrated to *.NET Core 3* project format
Migrated the `csproj` file to the new *Visual Studio* project format—which is a prerequisite for upgrading to *ASP.NET Core 3*. In addition, set the `<TargetFramework />` to `netcoreapp3.0` and the `<Project @Sdk />` to `Microsoft.NET.Sdk.Web`, which automatically configures *ASP.NET Core 3*. I maintained a number of elements from the legacy `csproj` file which aren't normally needed in the new *Visual Studio* project format, thus making it noisier than usual. These were highly-specific to particular features of the website, however, and thus I wanted to maintain them until we had confirmed that they weren't needed. This also includes updating the `packages.config` to the new project format. As part of this, I excluded a lot of `Microsoft` and `System` dependencies that are implicit in *ASP.NET Core 3*; I only included those needed to compile.