Excluded content folders from project
There are content folders, such as `Downloads`, `Help`, and `Media`, which are maintained by **GoldSim** and not part of the git repository. These may be maintained locally to keep parity with the live site, but should not be compiled during build, embedded into assemblies, or included in the **Visual Studio** file list.