Established `secrets.json` file
Previously, **GoldSim** relied on a `/Configuration/ConnectionString.config` file, which was `.gitignore`d to handle database connectivity. A better approach is to instead place thi within a `secrets.json` file. That said, an even better approach might be to use **Azure**'s new _Configuration_ capability; we will investigate this at a later date.