Introduce `Startup` class
In **ASP.NET Core** applications, the application is configured using a `Startup` class. This effectively replaces the `App_Start\RouteConfig` class used in **ASP.NET 2+ Framework** sites. The configuration is pretty much standard out-of-the-box defaults, with the exception of adding in support for **OnTopic** through e.g., `AddTopicSupport()`, `GoldSimActivator`, and `MapTopicRoute()`.