Introduced `CallsToActionViewComponent`
The `CallsToActionViewComponent` replaces the legacy `LayoutController.CallsToAction()` action by implementing it as a **View Component**. While much of the code appears new, it's actually just been moved around. The code in `CallsToActionViewComponent` is almost identical to the legacy `Layout.CallsToAction()` method body—the only difference is how the view is returned. Likewise, the `CallsToAction\Default.cshtml` is very similar to the legacy `CallsToAction.cshtml`—in this case, the main difference is the construction of the helper function, and the resulting indentation of the helper logic.