Introduced `FooterViewComponent`
The `FooterViewComponent` replaces the legacy `LayoutController.Footer()` action by implementing it as a **View Component**. As with the `CallsToActionViewComponent`, this is largely the same code as before, just shuffled around. The code in `FooterViewComponent` is almost identical to the legacy `Layout.Footer()` method body—the only difference is how the view is returned. Similarly, the only difference between `Footer\Default.cshtml` and the legacy `Footer.cshtml` is the namespace for the model.