Remove `WebFormsController` hack
Because the main site was previously in **ASP.NET MVC Framework**, but the forms were still using **ASP.NET WebForms**, a "bridge" was needed to allow the menu and footer to be displayed. With **ASP.NET Core** we'll no longer have access to **ASP.NET WebForms** (_oh, schucks_) so this will no longer be necessary. Good riddance!