Removed skins/themes
In **ASP.NET WebForms**, server and user control properties could be assigned default values based on `skin` files placed in the `App_Themes` directory. This was actually a pretty cool feature for theming display controls—and could be used for other functionality as well. That said, since we're no longer using **ASP.NET WebForms** this doesn't serve a purpose and can be deleted.