Remove legacy `Global.asax(.cs)`
With the migration to **ASP.NET Core** and the introduction of the `Startup` class, everything that was previously done by the `Global.asax(.cs)` file set is now handled in `Startup`. This includes configuring dependencies (via the `GoldSimActivator`), registering the _Composition Root_, registering the custom `TopicViewResultExecutor` (which replaced the legacy `TopicViewEngine`), registering routes (now via `UseEndpoints()`), and enforcing HTTPS (now using the new `UseHttpsRedirection()` middleware).