Migrated to `TopicView()` helper
In **ASP.NET MVC Framework**, the `ViewResult` (and, therefore, `TopicViewResult`) class was really simple to construct, requiring only a view model and view name (as well as a content type, in the `TopicViewResult`). In **ASP.NET Core**, the `ViewResult` requires a lot more information to be passed from the controller. This information is all available, but is tedious to reference each time. To mitigate that, the `TopicView()` helper was introduced—similar to the out-of-the-box `View()` helper.