Migrated `@Helper` to new function syntax
In **ASP.NET MVC Framework**, there was a useful `@helper` tag for encapulating repetitive HTML templates. This isn't supported in **ASP.NET Core**, though **ASP.NET Core 3** does allow something _similar_ through inline markup in display functions. In most cases, this is a simple matter of putting the code in a `@{}` block, setting the function type to `void`, and encapsulating the call in `@{}`. That said, these updates _appear_ much more invasive than they are because the nesting in a `@{}` results in indenting all of the functions by two spaces; the actual code hasn't really changed, though.