Migrated recursive `@helper` to `IHtmlContent`
While **ASP.NET Core** does allow inline HTML in display functions, it unfortunately doesn't (appear to) have a great solution for recursing over these. Given that, this is a bit of an ugly hack using the `IHtmlContent` approach. It should work, but I'd like to revisit it and see if we can't come up with a better approach later—such as, perhaps, just taking the time to put this in a reusable **View Component** that's part of the **OnTopic** library itself.