Removed header from `Response.Write()`
In the **ASP.NET MVC Framework**, placing the `<?xml?>` tag after the opening `@{}` code block resulted in it rendering with a blank line above it, which caused **Google** to reject it. To mitigate that, the `<?xml?>` element was written using `Response.Write()`. That method isn't supported via **Razor** files in **ASP.NET Core 3**, so I've put it outside of the codeblock. We'll need to evaluate if that actually works, or just reintroducses the original problem. If the latter, we'll need to revisit this—potentially by encoding the string and using `Context.Response.WriteAsync()`. Hopefully that's not required here.