Introduced `IConfiguration` dependency in `IControllerActivator`
There are arguments for and against this. Previously, instead of relaying the entire `Configuration` object, we just relayed the `GetConnectionString()` value. That said, services that require dynamic access to dependencies, such as `BraintreeConfiguration`, need access to the original `IConfiguration`, or an abstraction thereof. Given that, it makes sense to just relay the `IConfiguration` instead of just individual variables, such as the connection string. That said, this is a bit of a premature optimization; currently, we are configuring `BraintreeConfiguration` via the **OnTopic** library, so we can easily choose to revert this commit in the future if we need.