Introduced `IConfiguration` dependency in `BraintreeConfiguration`
Previously, `BraintreeConfiguration` relied on the static `ConfigurationManager` for ambient access to configuration variables defined in the `web.config`'s `<AppSettings />` element. That obviously isn't preferable, nor is it supported in **ASP.NET Core**. As such, this has been replaced with an explicitly injected dependency on `IConfiguration`, which is relayed from `Startup`. That said, our current configuration of `BraintreeConfiguration` is actually handled by **OnTopic**, so we may want to revisit the flexibility baked in at some point. The main advantage of this is that it gives us the _option_ to easily and temporarily override the values in a development enviroment. If this isn't needed, however, or if we're happy to just rely on e.g. `GetEnvironmentVariable()`, then we can do away with this dependency entirely. For now, we're keeping it for maintenance of previous functionality.