Removed extraneous `@{}` wrappers
In commit #670d1b2, I was overzealous in introducing `@{}` wrappers for calls to void functions. Many of these were already in blocks that **Razor** was evaluating as code—such as `for()` loops—and thus didn't need the `@` prefix at all (nevertheless the extraneous `{}`. Whoops! (These would be needed if they were in an HTML block.)