Clarified ambiguous `PageGroupTopicViewModel` reference
There are two `PageGroupTopicViewModel`s—the base class in `Ignia.Topics.ViewModels`, and a **GoldSim** specific version in `GoldSim.Web.Models`, which extends the original with e.g. `ThumbnailImage` and `WebPath`, thus allowing these to be treated as `ICardViewModel`s and looped against on index pages. This clarifies that ambiguity. (Though it's not entirely clear why this wasn't necessary in the previous version.)