Quick replace for `AntiXss`'s `Encoder.XmlEncode`
I was mistaken in commit #a54aa86—we _did_ still have a dependency on `AntiXss`. Whoops. Since we have a pretty predictable range of expected values to encode, this can be replaced with a simple series of replace statements. It's not an elegant solution, and we may want to revisit this later, but it should get the job done. (We could also use the built-in `SecurityElement.Escape()` method, but it's slow, and unnecessarily encodes quotes and apostrophes.)