Removed legacy `@RenderPage` call
Previously, in the **ASP.NET MVC Framework**, `@RenderPage` was necessary over e.g. `@Html.Partial()` for reasons that I'd need to look up. I believe it may have had to do with the context getting lost, and thus calls within `ContentList.cshtml` that depend on e.g. the query string not working. In absence of `@RenderPage` this will need to be reevaluated.