Migrated *Bing* configuration to `Secrets.json`
As a temporary fix, we're directly injecting the `IConfiguration` abstraction into `Search.cshtml` to replace the legacy `ConfigurationManaget`. This isn't elegant. A better approach would be to create a new `SearchViewModel` class with these setup as properties, and set them via either a custom controller and/or place them within **OnTopic** and use the standard `ITopicMappingService` to set them. For now, however, this helps get the app to a compiling state, at which point we can circle back on these issues.