Removed legacy sections
The `web.config` file is no longer being used, since it is an artifact of **ASP.NET Framework**. That said, not everything in it has been accounted for yet; some elements may need additional configuration to ensure that they continue to work in **ASP.NET Core**. For that reason, I'm keeping it around as a type of checklist. I've removed everything that's accounted for, and thus we don't need to worry about. We'll need to take a closer look at the elements that remain.