Moved CSS inside null condition
The `PageLevelNavigation` partial view is included on every page via the `_Layout.cshtml`. But not every page has a `PageLevelNavigation`. As such, it places the markup in a condition that checks for `Model.NavigationRoot`. Previously, the CSS was defined outside of the model. This is fine, since the element it decorates isn't visible; it has no impact on the user interface. That said, it adds unnecessary clutter (and weight) to every other page, and could potentially slow down the CSS processing by providing invalid CSS to contend with (since these definitions wouldn't actually include a valid `background-image` URL.