Ensured server-side commenting
With the refactoring of `Sitemap.cshtml` in commit #df2e620, I inadvertently missed that the server-side comments had been reinterpreted as client-side comments. Oops! This wasn't noticed by **Razor**'s syntax highlighting because it's valid HTML. But it's _not_ valid XML. I've updated this to use the `@*…*@` syntax for multi-line comments, instead of the previous `/*…*/` that's used for **C#**.