Introduced vendor dependencies
Previously, the `package.json` was exclusively used for managing dependencies of the **Gulp** workflow. This update additionally includes all of the vendor javascript and CSS files that we rely on. These will still just be installed into `/node_modules`, of course, so to fully support them we'll need to configure **Gulp** to copy them to their correct location in  `/wwwroot/Shared`; once that is done, we can remove the manually downloaded versions from **Git**. Until then, the manually downloaded versions are being retained in their current locations in `/Shared` and `wwwroot/Shared`.