Removed view model hack
In **ASP.NET Core 2.2**, dependency assemblies were not initialized automatically and, thus, made available to reflection for e.g., `ITopicMappingService` to discover. As such, they had to be "touched" in `Startup` to ensure that classes within those assemblies were discoverable. In **ASP.NET Core 3** this issue is resolved, so this workaround isn't required anymore.