Introduced `IsIndexed` property
On **GoldSim**, the `ContentList` content type is extended with an `IndexedList` view which displays a table of contents (via `_tableOfContents.cshtml`) at the top of the page with an achored index of each topic. Because this code is otherwise identical to `ContentList.cshtml`, the `IndexedList.cshtml` previously used **ASP.NET MVC Framework**'s `@RenderPage()` to include it—and, then, `_tableOfContents.cshtml` used the `View.Path` property to determine if it was an `IndexedList` and, if so, display the table of contents. This approach is problematic for a few reasons. But the biggest is because without `@RenderPage` support in **ASP.NET Core**, we must instead use `@Html.Partial()`—and that becomes the `View.Path`. To mitigate this, I've subtyped `ContentListTopicViewModel` for **GoldSim**, added an `IsIndexed` property, and explicitly set that on the `IndexedList.cshtml`. This also requires that the `@model` directive be updated on all `ContentList` views to disambiguate between `Ignia.Topics.ViewModels` and `GoldSim.Web.Models`.