Derived from `TopicController`
By deriving from `TopicController`—and introducing a dependency on `ITopicMappingService`—we gain the ability to map a form to an underlying topic, and thus import configuration settings and body text, as appropriate. Some text will likely continue to be built in—like labels or inline text—but giving **GoldSim** the ability to update the titles and introduction text should go a long way to minimizing the need for actual file edits.