Introduced `FormPageTopicViewModel<T>`
The `FormPageTopicViewModel<T>` is just a standard `PageTopicViewModel` with the addition of a `BindingModel` property of type `T`, which can be constructed using the built-in `FormsController.CreateViewModel()` factory method. This allows the form view to be bound to overall data—and to display the overall navigation!—while also including the binding model to use for form validation and model binding of values.