Removed default value for `Trainer`
Because `Trainer` is an _optional_ set of fields, but the `CoreContact` has `[Required]` attributes, data validation will (logically) flag those fields as missing if the `Trainer` property is instantiated. So long as it remains uninitialized (and, thus, `null`), however, those fields will remain optional (assuming the user doesn't fill out any of the fields).