Established `OptionalContact` as base class
This isn't the most elegant approach, but there are cases (such as `TrialFormBindingModel`) where a contact is optional, whereas it's required in most cases. To mimimize the amount of duplicated code, a new `OptionalContact` class was created under `CoreContact` which defines the main attributes, with `CoreContact` simply adding the `[Required]` attribute.