Introduced partial views for contacts
While having two separate classes for `OptionalContact` and `CoreContact` (see #328cda6) isn't as elegant as I'd prefer, having two separate views is downright ugly. That said, to keep them strongly typed, while also maintaining the client-side `[Required]` validation on `CoreContact`, this seems to be required. Fortunately it's a small class. This is related to issue #12.