Moved *Application Insights* configuration to secrets
Moved the **Application Insights** instrumentation key to the `secrets.json` file. Technically, it isn't necessarily a secret, _but_ doing so helps move toward keeping configuration in a predictable place—and, especially, this is also stored in an **Azure App Service** configuration key (which is where other items in `secrets.json` are stored on the server).