Removed `appSettings.config`
Previously, `<appSettings />` were stored in the `appSettings.config`. The application settings that remained were all default values from **ASP.NET MVC 5** and no longer necessary in **ASP.NET Core**. As such, it was removed. In the new site, such values are now stored in `secrets.json` for the most part.