Included `IISIntegration`
This application is intended to run on an **Azure App Service**, which will be running **Microsoft Internet Information Server (IIS)**. To accommodate a **.NET Core** application to run _in-process_ on **IIS**, a module is required. This is included on **Azure**, but for testing in **IISExpress** we are including this here. The `AspNetCoreModuleV3` is also configured via the `web.config`.