Generated changes from deployment to *IISExpress*
When running on **IISExpress**, or deploying to an **IIS** server, **ASP.NET Core** will automatically modify the `web.config` (if present). Technically, this doesn't _need_ to be committed, as it will be auto-generated—but, if it's not, the **ASP.NET Core** will add it each time the site is run, so it's easier just to keep it in the source instead of discarding the changes after every publish.