Introduced `StudentAcademic` form
The `StudentAcademic` form handles the request for an academic license for students. There remain some outstanding items with this with regard to, in particular, labels (e.g., "Name of institution" instead of "Organization Name"), replacing `Country` with `Department`, collecting sponsor `Telephone`, as well as the layout and labels of the sponsor section. In other words, there's a bit of work required here!