Introduced `AcademicSponsor` model, view
Originally, the _Student Sponsor_ section was simply modeled as a `CoreContact`. But neither the labels nor even the data model exactly match the `CoreContact` schema. Specifically, the labels are prefixed with "Sponsor", the "Organization" is labeled "Department", and, more importantly, it also requests a `PhoneNumber`. There may be ways around this, but without getting too cute it made more sense to simply subclass this. This introduces repetition in the views, but the markup is simple, so the cost is pretty trivial.