Introduced `_ExtendedContact` partial view
The `_ExtendedContact` provides a view for the `ExtendedContact` model, which extends `Contact` with an `Address`. Unfortunately, because the address is generally rendered between the `CoreContact` information (e.g., `FirstName`, `Email`, &c.) and the `Contact` information (e.g., `Country`, `PhoneNumber`), this can't simply reuse `_Contact`, and instead repeats the latter two. Those _could_ be centralized via an _additional_ partial, if we wanted, but it's a very small amount of simple markup so it's not an immediate priority.