Removed `OptionalContact` model and view
With the new ability to conditionally display and validate (groups of) fields, as introduced in #b4b3b00 and #0110d39, there is no longer a need for the duplication created by the `OptionalContact`. The `OptionalContact` model and the corresponding `_OptionalContact` partial had been previously introduced in #328cda6 and #794922b respectively as a hack to mitigate the fact that in some cases the contact is optional. This is more elegantly handled by making the parent properties nullable, and conditionally disabling the fields in the views. Good riddance!