Moved `Module` labels into form views
Previously, the label for `Modules` was hardcoded into the `_Modules.cshtml` as "I am also interested in:". This label, however, was only relevant to the `Trial` and `Demo` forms. For other forms, such as the `InstructorAcademic`, `StudentAcademic`, `Quote`, and `Purchase`, other labels are used. Decentralizing these does create some redundancy, but it also provides more flexibility.