Introduced `_Country`, `_PhoneNumber` partials
Previously, the markup for `Contact`'s `Country` and `PhoneNumber` properties were decentralized between the `_Contact` and `_ExtendedContact` partials, due to the fact that `_ExtendedContact` needed the address injected in between `_CoreContact` and `_Country`. To centralize these, a new `_Country` and `_PhoneNumber` partial have been introduced. That said, since `_Country` is sometimes full-width (`class="cell"`) and othertimes half width (`class="medium-6 cell"`) it needs special handling. To accommodate this, a `_ContactLocation` partial is introduced that composites the `_Country` and `_PhoneNumber` partials into a two-column format, while the `_ExtendedContact` calls them directly, as is. This isn't terribly elegant, and we may want to revisit this later by parameterizing this. For now, however, it works while centralizing the code. And while that may seem trivial currently, it will likely be a bigger issue once the county piece is filled in with actual countries.