Moved email instructions to `_Contact`
The email instructions are not used in references to `_ExtendedContact`. Since `_ExtendedContact` includes `_CoreContact`, this mistakenly meant that the email instructions were included on a number of forms they weren't intended, such as `Purchase` and `InstructorAcademic`. By moving this instead to `_Contact`, it only shows up on the relevant forms—namely, `Trial` and `Demo`.