Updated dependencies for `MenuViewComponent`
A feature update has been made to `MenuViewComponentBase<T>` in `Ignia.Topics.AspNetCore.Mvc` that necessitates that an `ITopicRepository` be passed to the constructor—and, thus, relayed through intermediaries such as `MenuViewComponent`. This update allows the menu navigation to be rerouted to a different root topic by setting a `NavigationRoot` attribute on the root container. This is useful for the forms, as it allows the navigation from `/Web` to be displayed, even though the form pages are under `/Forms`.