Introduced new `Form` attributes
In **OnTopic**, there is a `Form` content type descriptor that inherits from `Page`. This defines a number of attributes that map to content on the form pages. Using these gives **GoldSim** the ability to modify some common attributes, such as the button label, follow-up page, email address for the receipt, &c. Note: These are not yet integrated with the forms themselves; for now, they're just added to the view model so they'll be available for integration in a future commit.