Centralized `_Form.cshtml` layout
Established a `_Form.cshtml` for centralizing the layout of forms, created a `_ViewStart.cshtml` in the `/View/Forms` directory to set the `_Forms.cshtml` file as the layout, and then modifies each of the forms views to remove common elements that were now centralized. This looks like a bigger change than it actually is because the form bodies were unindented to accommodate the reduced nesting caused by the above centralization.