Configured deployment rules
Generally, we set our deployment to leave files on the server, instead of doing a full synchronization. Nevertheless, to be safe, there are directories that are not locally maintained on development machines, nor checked into source control, and should thus be actively excluded from any deployment in order to avoid updating or deleting custom deployed content. The use of the `<MsDeploySkipRules />` has the advantage of not deleting server-side contnet.