Established new *npm* project
After upgrading to **Gulp 4**, I experienced quite a few conflicts with dependencies. In addition, the previous `package.json` was based on the `Gulp-Workflow` project, and not on a **GoldSim** specific version, which causes confusion. To establish a clean slate, a new `package.json` file has been established. Dependencies will be reintroduced as appropriate.