Implemented `fontawesome-free` via *npm*
Previously, **Font Awesome** was manually downloaded and included as part of the **GoldSim** repository. It is now referenced via **npm**, included directly via the `_fonts.scss`, with dependencies copied over via **Gulp**. Because this is a newer version of **Font Awesome**, some class references needed to be updated; importantly, references to _brands_ are now in a separate font file and included via `fab` not `fa`.