Remove *Foundation* TypeScript compilation
As of commit #3804c77, the **Foundation** JavaScript files are being sourced directly from the pre-compiled version in `node_modules` and, thus, no longer need to build the `foundation.d.ts` file. We'll want to migrate **GoldSim** over to TypeScript at some point, and will need to revisit the best practive for compiling it (e.g., via **Gulp** or **MSBuild**), but we'll defer that question until it's actually needed.