Added `SaveAsTopic` property
Some `FormPage`s are intended to be saved as a topic to the database. Not all require this feature, however. To keep the form processing general, while giving content editor more control over functionality, this is exposed as a checkbox on the `TopicPage` content type, and similarly exposed as a `SaveAsTopic` property on the `FormPageTopicViewModel`. While I was at it, reformatted the `ProcessForm()` method in `FormsController` to better break up distinct code blocks.