Defined `$dropdownmenu-background` as transparent
It seems that **Foundation 6** has introduced a new `$dropdownmenu-background` variable—used for the background color of menu links—which defaults to #FFFFFF (white). Not only is that not the correct background color for us, but we also want the background of the menu itself to be ued here, which varies based on tier. Given that, I reset this to transparent. For this to take effect, this also required moving the `foundationSettings` above the `variables`.