Updated *OnTopic* to `3.8.0-csharp8-update.1+357` package.
Unfortunately, there remains an issue with **NuGet** not supporting **SemVer 2.0** and, thus, not recognizing the `+357` (commit count) on the **NuGet** package names. As such, there will likely remain issues in the future with downloading these packages. That said, this at least ensures that this build is stable against **NuGet** for the current version—and, hopefully, will remain so until we are ready to deploy to `develop`, at least, thus reducing the need to sideload dependencies unless there's a new feature or hotfix on the `csharp-update` branch.