Remove initialization pages
Page initialization was a previous shortcut used to ensure just-in-time compilation occurred immediately after the application restarted, to prevent slow rendering times for the first user to hit each unique view. It's not entirely clear that this worked as advertised. Regardless, with support compiled views in **ASP.NET Core 3** (hooray!) this no longer serves a purpose.