Introduced `gulp js` task for JavaScript files
Currently, the JavaScript build process is pretty simple: The files are put through `JSHint`, then `Uglify`ed, `concat`enates the files, and finally copied to the build folder (`/wwwroot/Shared/Scripts/`). It then does the same thing for the views folder, but skips `concat`. Also updated the location of view-specific scripts to be in the `/Views` folder.