Enable support for `NoIndex` attribute
This is a hotfix to prevent pages from being indexed. In the new `DotNetCore` release we'll formalize this into the actual `_layout.cshtml`, but that requires a more invasive update to the view models. In the short-term, this allows, at minimum, individual page templates (such as `Payments.cshtml`) to be hidden from search engines, since they're only intended to be accessed by request.