Removed (now duplicate) `WebPath` property
The `WebPath` is now implemented on the base `TopicViewModel` class, so it doesn't need to be reintroduced on view models that don't derive from `PageTopicViewModel` (where it was previously implemented). This addresses the fact that while only web pages may be *intended* to be accessed by web URLs, there are other view models and content types that may *have* a `WebPath` (e.g., for JSON endpoints).